Convert Log int/format printing methods to functions

+++ interpreter/javascript/eon/Any.c
@@ -6,7 +6,7 @@ union e_Any e_Any_ptr(void* ptr) {
     return any;
 }
 
-union e_Any e_Any_val(intptr_t val) {
+union e_Any e_Any_val(unsigned long val) {
     union e_Any any;
     any.val = val;
     return any;
